pyproject.toml
[project]
name = "nemo"
version = "0.1.0"
description = "Transcription and diarization toolkit using Whisper and NeMo. Originally written for use on Exeter's ISCA server, however it can now be built on any platform."
authors = [
    {name = "Adam El Kholy"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "=3.11.3"
dependencies = [
    "absl-py (==2.3.1)",
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.12.15)",
    "aiosignal (==1.4.0)",
    "alembic (==1.16.4)",
    "annotated-types (==0.7.0)",
    "antlr4-python3-runtime (==4.9.3)",
    "asteroid-filterbanks (==0.4.0)",
    "asttokens (==3.0.0)",
    "attrs (==25.3.0)",
    "audioread (==3.0.1)",
    "av (==15.0.0)",
    "blis (==1.2.1)",
    "braceexpand (==0.1.7)",
    "catalogue (==2.0.10)",
    "certifi (==2025.8.3)",
    "cffi (==1.17.1)",
    "charset-normalizer (==3.4.3)",
    "click (==8.2.1)",
    "cloudpathlib (==0.21.1)",
    "cloudpickle (==3.1.1)",
    "colorama (==0.4.6)",
    "coloredlogs (==15.0.1)",
    "colorlog (==6.9.0)",
    "comm (==0.2.3)",
    "confection (==0.1.5)",
    "contourpy (==1.3.3)",
    "cryptography (==44.0.3)",
    "ctranslate2 (==4.4.0)",
    "cycler (==0.12.1)",
    "cymem (==2.0.11)",
    "cytoolz (==1.0.1)",
    "datasets (==3.2.0)",
    "decorator (==5.2.1)",
    "dill (==0.3.8)",
    "distance (==0.1.3)",
    "docopt (==0.6.2)",
    "dora-search (==0.1.12)",
    "editdistance (==0.8.1)",
    "einops (==0.8.1)",
    "executing (==2.2.0)",
    "faster-whisper (==1.1.0)",
    "fiddle (==0.3.0)",
    "filelock (==3.19.1)",
    "flatbuffers (==25.2.10)",
    "fonttools (==4.59.1)",
    "frozenlist (==1.7.0)",
    "fsspec[http] (==2024.9.0)",
    "future (==1.0.0)",
    "g2p-en (==2.1.0)",
    "gitdb (==4.0.12)",
    "gitpython (==3.1.45)",
    "graphviz (==0.21)",
    "greenlet (==3.2.4)",
    "grpcio (==1.74.0)",
    "huggingface-hub (==0.23.5)",
    "humanfriendly (==10.0)",
    "hydra-core (==1.3.2)",
    "hyperpyyaml (==1.2.2)",
    "idna (==3.10)",
    "inflect (==7.5.0)",
    "iniconfig (==2.1.0)",
    "intervaltree (==3.1.0)",
    "ipython (==9.4.0)",
    "ipython-pygments-lexers (==1.1.1)",
    "ipywidgets (==8.1.7)",
    "jedi (==0.19.2)",
    "jinja2 (==3.1.6)",
    "jiwer (==4.0.0)",
    "joblib (==1.5.2)",
    "julius (==0.2.7)",
    "jupyterlab-widgets (==3.0.15)",
    "kaldi-python-io (==1.2.2)",
    "kaldiio (==2.18.1)",
    "kiwisolver (==1.4.9)",
    "lameenc (==1.8.1)",
    "langcodes (==3.5.0)",
    "language-data (==1.3.0)",
    "lazy-loader (==0.4)",
    "levenshtein (==0.27.1)",
    "lhotse (==1.30.3)",
    "libcst (==1.8.2)",
    "librosa (==0.11.0)",
    "lightning (==2.5.3)",
    "lightning-utilities (==0.15.2)",
    "lilcom (==1.8.1)",
    "llvmlite (==0.44.0)",
    "loguru (==0.7.3)",
    "mako (==1.3.10)",
    "marisa-trie (==1.3.1)",
    "markdown (==3.8.2)",
    "markdown-it-py (==4.0.0)",
    "markupsafe (==3.0.2)",
    "marshmallow (==4.0.0)",
    "matplotlib (==3.10.5)",
    "matplotlib-inline (==0.1.7)",
    "mdurl (==0.1.2)",
    "ml-dtypes (==0.5.3)",
    "more-itertools (==10.7.0)",
    "mpmath (==1.3.0)",
    "msgpack (==1.1.1)",
    "multidict (==6.6.4)",
    "multiprocess (==0.70.16)",
    "murmurhash (==1.0.13)",
    "nemo-toolkit[asr] (==2.0.0rc0)",
    "networkx (==3.5)",
    "nltk (==3.9.1)",
    "numba (==0.61.2)",
    "numpy (==1.26.4)",
    "omegaconf (==2.3.0)",
    "onnx (==1.19.0)",
    "onnxruntime (==1.22.1)",
    "openunmix (==1.3.0)",
    "optuna (==4.5.0)",
    "packaging (==25.0)",
    "pandas (==2.3.2)",
    "parso (==0.8.5)",
    "phonenumbers (==9.0.12)",
    "pillow (==11.3.0)",
    "plac (==1.4.5)",
    "platformdirs (==4.4.0)",
    "pluggy (==1.6.0)",
    "pooch (==1.8.2)",
    "preshed (==3.0.10)",
    "presidio-analyzer (==2.2.359)",
    "presidio-anonymizer (==2.2.359)",
    "primepy (==1.3)",
    "prompt-toolkit (==3.0.51)",
    "propcache (==0.3.2)",
    "protobuf (==6.32.0)",
    "pure-eval (==0.2.3)",
    "pyannote-audio (==3.3.2)",
    "pyannote-core (==5.0.0)",
    "pyannote-database (==5.1.3)",
    "pyannote-metrics (==3.2.1)",
    "pyannote-pipeline (==3.0.1)",
    "pyarrow (==21.0.0)",
    "pybind11 (==3.0.1)",
    "pycparser (==2.22)",
    "pydantic (==2.11.7)",
    "pydantic-core (==2.33.2)",
    "pydub (==0.25.1)",
    "pygments (==2.19.2)",
    "pyloudnorm (==0.1.1)",
    "pyparsing (==3.2.3)",
    "pyreadline3 (==3.5.4)",
    "pytest (==8.4.1)",
    "python-dateutil (==2.9.0.post0)",
    "pytorch-lightning (==2.5.3)",
    "pytorch-metric-learning (==2.9.0)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.2)",
    "rapidfuzz (==3.13.0)",
    "regex (==2025.7.34)",
    "requests (==2.32.5)",
    "requests-file (==2.1.0)",
    "resampy (==0.4.3)",
    "retrying (==1.4.2)",
    "rich (==14.1.0)",
    "ruamel-yaml (==0.18.15)",
    "ruamel-yaml-clib (==0.2.12)",
    "sacremoses (==0.1.1)",
    "safetensors (==0.6.2)",
    "scikit-learn (==1.7.1)",
    "scipy (==1.16.1)",
    "semver (==3.0.4)",
    "sentencepiece (==0.2.1)",
    "sentry-sdk (==2.35.1)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "smart-open (==7.3.0.post1)",
    "smmap (==5.0.2)",
    "sortedcontainers (==2.4.0)",
    "soundfile (==0.13.1)",
    "sox (==1.5.0)",
    "soxr (==0.5.0.post1)",
    "spacy (==3.8.7)",
    "spacy-legacy (==3.0.12)",
    "spacy-loggers (==1.0.5)",
    "speechbrain (==1.0.3)",
    "sqlalchemy (==2.0.43)",
    "srsly (==2.5.1)",
    "stack-data (==0.6.3)",
    "submitit (==1.5.3)",
    "sympy (==1.14.0)",
    "tabulate (==0.9.0)",
    "tensorboard (==2.20.0)",
    "tensorboard-data-server (==0.7.2)",
    "tensorboardx (==2.6.4)",
    "termcolor (==3.1.0)",
    "text-unidecode (==1.3)",
    "texterrors (==1.0.9)",
    "thinc (==8.3.4)",
    "threadpoolctl (==3.6.0)",
    "tldextract (==5.3.0)",
    "tokenizers (==0.19.1)",
    "toolz (==1.0.0)",
    "torch (==2.8.0)",
    "torch-audiomentations (==0.12.0)",
    "torch-pitch-shift (==1.2.5)",
    "torchaudio (==2.8.0)",
    "torchmetrics (==1.8.1)",
    "tqdm (==4.67.1)",
    "traitlets (==5.14.3)",
    "transformers (==4.40.2)",
    "treetable (==0.2.5)",
    "typeguard (==4.4.4)",
    "typer (==0.16.1)",
    "typing-extensions (==4.15.0)",
    "typing-inspection (==0.4.1)",
    "tzdata (==2025.2)",
    "unidecode (==1.4.0)",
    "urllib3 (==2.5.0)",
    "uroman (==1.3.1.1)",
    "wandb (==0.21.1)",
    "wasabi (==1.1.3)",
    "wcwidth (==0.2.13)",
    "weasel (==0.4.1)",
    "webdataset (==1.0.2)",
    "werkzeug (==3.1.3)",
    "wget (==3.2)",
    "whisperx (==3.3.1)",
    "widgetsnbextension (==4.0.14)",
    "win32-setctime (==1.2.0)",
    "wrapt (==1.17.3)",
    "xxhash (==3.5.0)",
    "yarl (==1.20.1)",
    "ctc-forced-aligner @ git+ssh://git@github.com/MahmoudAshraf97/ctc-forced-aligner.git",
    "deepmultilingualpunctuation @ git+ssh://git@github.com/oliverguhr/deepmultilingualpunctuation.git",
    "demucs @ git+ssh://git@github.com/MahmoudAshraf97/demucs.git"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false